PLONEOPS-17452 : Updated Managed services request and response.
