--- src/main/java/com/oneops/proxy/model/ms/Credentials.java
@@ -1,19 +1,0 @@
-package com.oneops.proxy.model.ms;
-
-/**
- * Object which is used with in Managed service Auth API with "MSClientAuthRequest"
- *
- * @author Varsha
- */
-public class Credentials {
-
-  private String user;
-
-  public String getUser() {
-    return user;
-  }
-
-  public void setUser(String user) {
-    this.user = user;
-  }
-}
