+++ src/main/java/com/oneops/proxy/model/ms/MSClientAuthResponse.java
@@ -1,6 +1,6 @@
 package com.oneops.proxy.model.ms;
 
-import java.util.List;
+import java.util.Map;
 
 /**
  * Response object for Auth Api in Managed Services. Return if true/false if user is authorized with
@@ -12,7 +12,10 @@ public class MSClientAuthResponse {
 
   private String namespace;
 
-  private List<AuthorizedUser> authorized;
+  private String user;
+  private boolean authorized;
+  private String error;
+  private Map<String, String> metadata;
 
   public String getNamespace() {
     return namespace;
@@ -22,11 +25,35 @@ public class MSClientAuthResponse {
     this.namespace = namespace;
   }
 
-  public List<AuthorizedUser> getAuthorized() {
+  public String getUser() {
+    return user;
+  }
+
+  public void setUser(String user) {
+    this.user = user;
+  }
+
+  public boolean isAuthorized() {
     return authorized;
   }
 
-  public void setAuthorized(List<AuthorizedUser> authorized) {
+  public void setAuthorized(boolean authorized) {
     this.authorized = authorized;
+  }
+
+  public String getError() {
+    return error;
+  }
+
+  public void setError(String error) {
+    this.error = error;
+  }
+
+  public Map<String, String> getMetadata() {
+    return metadata;
+  }
+
+  public void setMetadata(Map<String, String> metadata) {
+    this.metadata = metadata;
   }
 }
